remove some warnings

+++ voldor/geometry.cpp
@@ -307,9 +307,9 @@ solve_p3p_pool
 	//case 5:  poses_pool_used = batch_cpu_solver_tft(trifocal_0_map.data(), trifocal_1_map.data(), trifocal_2_map.data(), trifocal_0_map.size(), cams[active_idx].K, cfg.n_poses_to_sample, (float*)poses_pool.data, (float*)next_pool->data(), cfg.batch_workers, false); break;
 
 	// gpm
-	case  8: poses_pool_used = batch_cpu_solver_gpm(trifocal_0_map.data(), trifocal_1_map.data(), trifocal_0_map.size(), cams[active_idx].K, 0, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false); break;
-	case  9: poses_pool_used = batch_cpu_solver_gpm(trifocal_0_map.data(), trifocal_1_map.data(), trifocal_0_map.size(), cams[active_idx].K, 1, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false); break;
-	case 10: poses_pool_used = batch_cpu_solver_gpm(trifocal_0_map.data(), trifocal_1_map.data(), trifocal_0_map.size(), cams[active_idx].K, 2, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false); break;
+	case  8: poses_pool_used = batch_cpu_solver_gpm(trifocal_0_map.data(), trifocal_1_map.data(), (int)trifocal_0_map.size(), cams[active_idx].K, 0, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false); break;
+	case  9: poses_pool_used = batch_cpu_solver_gpm(trifocal_0_map.data(), trifocal_1_map.data(), (int)trifocal_0_map.size(), cams[active_idx].K, 1, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false); break;
+	case 10: poses_pool_used = batch_cpu_solver_gpm(trifocal_0_map.data(), trifocal_1_map.data(), (int)trifocal_0_map.size(), cams[active_idx].K, 2, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false); break;
 
 	// rnp
 	case 16: poses_pool_used = batch_cpu_solver_rnp(pts3_map.data(), pts2_map.data(), (int)pts2_map.size(), cams[active_idx].K, 0, cfg.n_poses_to_sample, (float*)poses_pool.data, cfg.batch_workers, false, cfg.rs_direction, cfg.rs_r0, cfg.rs_max_iterations); break;
